Добавил политику пользователя

+++ cms/app/Http/Controllers/CouponController.php
@@ -2,20 +2,45 @@
 
 namespace CMS\Http\Controllers;
 
+use Gate;
+use Validator;
 use CMS\Coupon;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CouponController extends Controller
 {
+    /**
+     * Undocumented function
+     *
+     * @return void
+     */
     public function add(){
-
         return view('coupon.add');
-
     }
-
+    /**
+     * Undocumented function
+     *
+     * @param Request $request
+     * @return void
+     */
     public function create(Request $request){
 
+        if (Gate::allow('administrate', Auth::user())) {
+            abort(403, 'У вас нет прав на редактирование данного материала');
+        }
+        $validator = Validator::make($request->all(), [
+            'name' => 'required|max:255',
+            'procent' => 'required',
+            'date' => 'required|date',
+            'status' => 'required',
+        ]);
+        if ($category->fails()) {
+            return redirect('coupon/add')
+                        ->withErrors($validator)
+                        ->withInput();
+        }
         Coupon::create([
             'name' => $request->couponName,
             'procent' => $request->procentCoupon,
@@ -23,9 +48,13 @@ class CouponController extends Controller
             'status' => 'active'
         ]);
 
-        return redirect()->route('CouponAdd');
+        return redirect()->route('CouponAll');
     }
-
+    /**
+     * Undocumented function
+     *
+     * @return void
+     */
     public function all(){
         
         return view('coupon.all', [
@@ -33,9 +62,17 @@ class CouponController extends Controller
         ]);
 
     }
-
+    /**
+     * Undocumented function
+     *
+     * @param [type] $id
+     * @return void
+     */
     public function delete($id){
 
+        if (Gate::allow('administrate', Auth::user())) {
+            abort(403, 'У вас нет прав на редактирование данного материала');
+        }
         Coupon::findOrFail($id)->delete();
 
         return redirect()->route('CouponAll');
